+++ .eslintrc.js
@@ -24,5 +24,17 @@ module.exports = {
                 'cypress/globals': true,
             },
         },
+        {
+            files: ['**/packages/docs/**/*.vue'],
+            plugins: ['vue'],
+            extends: ['plugin:vue/vue3-recommended'],
+            parserOptions: {
+                ecmaVersion: 2020,
+                parser: '@typescript-eslint/parser',
+            },
+            rules: {
+                'vue/html-indent': ['warn', 4],
+            },
+        },
     ],
 };