+++ packages/core/src/primitives/int.ts
@@ -7,6 +7,9 @@ import { assert } from '@scale-codec/util';
  */
 export type Endianness = 'le' | 'be';
 
+/**
+ * Bits count supported by integer codecs
+ */
 export type AllowedBits = 8 | 16 | 32 | 64 | 128; // and so on, but let's be realistic
 
 export interface BigIntCodecOptions {
@@ -59,6 +62,10 @@ class BytesRepr {
 }
 
 export function encodeBigInt(value: JSBI, { bits, signed, endianness }: BigIntCodecOptions): Uint8Array {
+    if (!signed && JSBI.lessThan(value, JSBI.BigInt(0))) {
+        throw new Error('signed=false, but negative num is passed');
+    }
+
     const repr = new BytesRepr(bits, endianness);
 
     const truncated = signed ? JSBI.asIntN(bits, value) : JSBI.asUintN(bits, value);