+++ packages/definition-compiler/src/render/render.ts
@@ -10,16 +10,27 @@ import Fields from './components/Fields';
 import UseConfig from './components/UseConfig';
 import VoidAlias from './components/VoidAlias';
 
+export interface RenderNamespaceDefinitionParams {
+    /**
+     * Runtime library with STD codecs + reexports from the core library.
+     *
+     * @remarks
+     * See `@scale-codec/definition-runtime` library - it exists specially for it.
+     */
+    importLib: string;
+    /**
+     * Single tuples are always an arrays with a single element. It is possible to make final code cleaner
+     * (and a bit performant) if render such tuples just as aliases for the inner element. It is optional feature.
+     */
+    rollupSingleTuplesIntoAliases?: boolean;
+}
+
+/**
+ * Renders provided definition into a valid TypeScript code.
+ */
 export async function renderNamespaceDefinition(
     def: NamespaceDefinition,
-    params: {
-        importLib: string;
-        /**
-         * Single tuples are always an arrays with a single element. It is possible to make final code cleaner
-         * (and a bit performant) if render such tuples just as aliases for the inner element. It is optional feature.
-         */
-        rollupSingleTuplesIntoAliases?: boolean;
-    },
+    params: RenderNamespaceDefinitionParams,
 ): Promise<string> {
     const app = createApp(App, {
         defmap: def,