+++ packages/docs/windi.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'windicss/helpers';
+
+export default defineConfig({
+    preflight: false,
+    extract: {
+        include: ['**/*.{md,vue}', '.vitepress/**/*.{ts,md,vue}'],
+    },
+});