+++ packages/definition-compiler/CHANGELOG.md
@@ -1,12 +1,23 @@
 # @scale-codec/definition-compiler
 
-## 0.2.0
+## 0.3.0
+
 ### Minor Changes
 
-- ac55aa8: feat: new optional feature - rollup single tuple into an alias of the inner type
-- a1b1a98: feat: support definitions for external types
-- 4a77d16: feat: now it is possible to define codec aliases
+-   1904251: Update type exports and documentation
 
 ### Patch Changes
 
-- cfec796: Update dependencies
+-   07d0d6e: Hide `WithTMark` from public API; fix doc comment
+
+## 0.2.0
+
+### Minor Changes
+
+-   ac55aa8: feat: new optional feature - rollup single tuple into an alias of the inner type
+-   a1b1a98: feat: support definitions for external types
+-   4a77d16: feat: now it is possible to define codec aliases
+
+### Patch Changes
+
+-   cfec796: Update dependencies